scripts/requirements.txt
Faker
kafka-python
pyspark
psycopg2-binary